feat(shell): add commands for db utilities
There is a number of commands for new applets I created
for handling some personal info I would like to maintain
for a variety of projects.

These applets are not part of AUTOMATE and are kept in their own
projects, but I have created references here as a way to easily
add them to PATH and remove the .py extension.